backend-python/requirements.txt
boto3
openai
python-dotenv
flask
Flask==2.3.2
python-dotenv==1.0.0
requests==2.31.0
# Add any additional dependencies here
